+++ compiler/condsyms.nim
@@ -135,4 +135,5 @@ proc initDefines*(symbols: StringTableRef) =
   defineSymbol("nimHasIterable")
   defineSymbol("nimHasTypeofVoid")
   defineSymbol("nimHasDragonBox")
+  defineSymbol("nimHasDeferImport")
   defineSymbol("nimHasHintAll")