+++ tests/stdlib/mimportutils3.nim
@@ -1,0 +1,5 @@
+var count = 0
+
+proc mimportutils3_fn1*(): int {.exportc.} =
+  count.inc
+  count