+++ tests/stdlib/mimportutils_cyclic2.nim
@@ -1,0 +1,4 @@
+import mimportutils_cyclic1
+
+proc mimportutils_cyclic2_lib(): Foo {.exportc.} =
+  Foo(a: 1)