+++ changelog.md
@@ -379,6 +379,8 @@
 - Added a new module `std/importutils`, and an API `privateAccess`, which allows access to private fields
   for an object type in the current scope.
 
+- Added `importutils.deferImport` to mitigate cyclic import limitations.
+
 - `typeof(voidStmt)` now works and returns `void`.